+++ packages/typespec-go/test/armlargeinstance/zz_options.go
@@ -10,9 +10,6 @@ type AzureLargeInstanceClientBeginRestartOptions struct {
 	// Resumes the long-running operation from the provided token.
 	ResumeToken string
 
-	// Body parameter's content type. Known values are application/json
-	ContentType *string
-
 	// When set to 'active', this parameter empowers the server with the ability to forcefully terminate and halt any existing
 	// processes that may be running on the server
 	ForceParameter *ForceState