+++ packages/typespec-go/test/cadlranch/parameters/basicparamsgroup/zz_basicimplicitbody_client.go
@@ -50,12 +50,12 @@ func (client *BasicImplicitBodyClient) simpleCreateRequest(ctx context.Context, 
 	if err != nil {
 		return nil, err
 	}
-	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	body := struct {
 		Name string `json:"name"`
 	}{
 		Name: name,
 	}
+	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
 		return nil, err
 	}