+++ packages/typespec-go/test/cadlranch/parameters/bodyoptionalgroup/zz_options.go
@@ -20,16 +20,10 @@ type BodyOptionalityClientRequiredImplicitOptions struct {
 // method.
 type BodyOptionalityOptionalExplicitClientOmitOptions struct {
 	Body *BodyModel
-
-	// Body parameter's content type. Known values are application/json
-	ContentType *string
 }
 
 // BodyOptionalityOptionalExplicitClientSetOptions contains the optional parameters for the BodyOptionalityOptionalExplicitClient.Set
 // method.
 type BodyOptionalityOptionalExplicitClientSetOptions struct {
 	Body *BodyModel
-
-	// Body parameter's content type. Known values are application/json
-	ContentType *string
 }